Removed bad method from BehaviourContainer.cs
Splitted the DefaultLinkComponent.razor into 3 different ones.
Simplified the draw link options.
Started beautifying the sandbox project.
Zoom is now decimal.

+++ Blazored.Diagrams.Test/Diagrams/DiagramTests.cs
@@ -26,7 +26,7 @@ public class DiagramTests
     public void Test_Set_Zoom()
     {
         //Arrange
-        const double zoom = 2;
+        const decimal zoom = 2;
         var diagram = Instance;
 
         //Act
