+++ Blazored.Diagrams/Interfaces/IZoomable.cs
@@ -8,11 +8,11 @@ public interface IZoomable
     /// <summary>
     ///     Current zoom value
     /// </summary>
-    public double Zoom { get; set; }
+    public decimal Zoom { get; set; }
 
     /// <summary>
     ///     Sets the zoom to the specified value.
     /// </summary>
     /// <param name="zoom"></param>
-    public void SetZoom(double zoom);
+    public void SetZoom(decimal zoom);
 }
